trabajo_final/requirements.txt
# requirements.txt

pandas
python-dotenv

# LangChain
langchain
langchain-core
langchain-community
langchain-google-genai

# LangGraph
langgraph

# Chroma vectorstore
chromadb
langchain-chroma

# Notion API
notion-client

# LangSmith
langsmith

# Utilidades
typing-extensions